+++ Minecraft Exercise/ZTCraft/world/World.cpp
@@ -1,11 +1,11 @@
 #include "World.h"
 
-void World::render(double delta) {
+void World::render(float delta) {
 	 for (unsigned int i = 0; i < blocks.size(); ++i) {
 		  blocks[i]->update(delta);
 		  blocks[i]->render(delta);
 	 }
 }
 
-void World::update(double delta) {
+void World::update(float delta) {
 }
